KOGITO-5644 KOGITO-5645: add working spring boot example
