--- process-monitoring-quarkus/operator/process-quarkus-example.yaml
@@ -1,34 +1,0 @@
-apiVersion: app.kiegroup.org/v1beta1
-kind: KogitoBuild
-metadata:
-  name: process-quarkus-example
-spec:
-  type: RemoteSource
-  gitSource:
-    contextDir: process-quarkus-example
-    uri: 'https://github.com/kiegroup/kogito-examples'
-  # set your maven nexus repository to speed up the build time
-  #mavenMirrorURL:
-#uncomment to enable persistence
-#---
-##Infinispan operator should be pre-installed in namespace
-#apiVersion: app.kiegroup.org/v1beta1
-#kind: KogitoInfra
-#metadata:
-#  name: kogito-infinispan-infra
-#spec:
-#  resource:
-#    apiVersion: infinispan.org/v1
-#    kind: Infinispan
----
-apiVersion: app.kiegroup.org/v1beta1
-kind: KogitoRuntime
-metadata:
-  name: process-quarkus-example
-# uncomment to enable persistence
-#spec:
-#  infra:
-#    - kogito-infinispan-infra
-  #env:
-  #  - name: MAVEN_ARGS_APPEND
-  #    value: "-Ppersistence"
