+++ process-monitoring-quarkus/src/main/resources/application.properties
@@ -1,24 +1,2 @@
-# Packaging
-# quarkus.package.type=fast-jar
-
-quarkus.swagger-ui.always-include=true
-
-quarkus.infinispan-client.server-list=localhost:11222
-quarkus.http.cors=true
-
-kafka.bootstrap.servers=localhost:9092
-
-mp.messaging.outgoing.kogito-processinstances-events.connector=smallrye-kafka
-mp.messaging.outgoing.kogito-processinstances-events.topic=kogito-processinstances-events
-mp.messaging.outgoing.kogito-processinstances-events.value.serializer=org.apache.kafka.common.serialization.StringSerializer
-
-mp.messaging.outgoing.kogito-usertaskinstances-events.connector=smallrye-kafka
-mp.messaging.outgoing.kogito-usertaskinstances-events.topic=kogito-usertaskinstances-events
-mp.messaging.outgoing.kogito-usertaskinstances-events.value.serializer=org.apache.kafka.common.serialization.StringSerializer
-
-mp.messaging.outgoing.kogito-variables-events.connector=smallrye-kafka
-mp.messaging.outgoing.kogito-variables-events.topic=kogito-variables-events
-mp.messaging.outgoing.kogito-variables-events.value.serializer=org.apache.kafka.common.serialization.StringSerializer
-
 # Maximum Java heap to be used during the native image generation
 quarkus.native.native-image-xmx=6g
