+++ process-monitoring-springboot/Dockerfile
@@ -1,0 +1,5 @@
+FROM quay.io/kiegroup/kogito-runtime-jvm:latest
+
+ENV RUNTIME_TYPE springboot
+
+COPY target/process-monitoring-springboot.jar $KOGITO_HOME/bin/
