+++ process-monitoring-springboot/docker-compose/grafana/grafana.ini
@@ -1,0 +1,3 @@
+[auth.anonymous]
+enabled = true
+org_role = Admin
