+++ process-monitoring-springboot/docker-compose/grafana/provisioning/dashboards/dashboard.yml
@@ -1,0 +1,13 @@
+apiVersion: 1
+
+providers:
+- name: 'Prometheus'
+  orgId: 1
+  folder: ''
+  type: file
+  disableDeletion: true
+  allowUiUpdates: true
+  editable: true
+  options:
+    path: /etc/grafana/provisioning/dashboards
+
