--- process-monitoring-springboot/operator/process-springboot-example.yaml
@@ -1,40 +1,0 @@
-# uncomment to enable persistence
-##Infinispan operator should be pre-installed in namespace
-#apiVersion: app.kiegroup.org/v1beta1
-#kind: KogitoInfra
-#metadata:
-#  name: kogito-infinispan-infra
-#spec:
-#  resource:
-#    apiVersion: infinispan.org/v1
-#    kind: Infinispan
----
-apiVersion: app.kiegroup.org/v1beta1
-kind: KogitoBuild
-metadata:
-  name: process-springboot-example
-spec:
-  type: RemoteSource
-  runtime: springboot
-  #env:
-  # env can be used to set variables during build
-  #- name: MY_CUSTOM_ENV
-  #  value: "my value"
-  gitSource:
-    contextDir: process-springboot-example
-    uri: 'https://github.com/kiegroup/kogito-examples'
-  # set your maven nexus repository to speed up the build time
-  #mavenMirrorURL:
----
-apiVersion: app.kiegroup.org/v1beta1
-kind: KogitoRuntime
-metadata:
-  name: process-springboot-example
-spec:
-  runtime: springboot
-  # uncomment to enable persistence
-  #env:
-  #  - name: MAVEN_ARGS_APPEND
-  #    value: "-Ppersistence"
-#  infra:
-#    - kogito-infinispan-infra
