--- process-monitoring-springboot/src/main/java/org/kie/kogito/examples/PersonValidationService.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.examples;
-
-import org.kie.kogito.examples.demo.Person;
-import org.kie.kogito.rules.DataSource;
-import org.kie.kogito.rules.DataStore;
-import org.kie.kogito.rules.RuleUnitData;
-
-public class PersonValidationService implements RuleUnitData {
-    private DataStore<Person> persons = DataSource.createStore();
-
-    public DataStore<Person> getPersons() {
-        return persons;
-    }
-}
