--- process-monitoring-springboot/src/main/java/org/kie/kogito/examples/demo/Person.java
@@ -1,67 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.examples.demo;
-
-import java.io.Serializable;
-
-public class Person implements Serializable {
-
-    private static final long serialVersionUID = -571683427125356701L;
-
-    private String name;
-    private int age;
-    private boolean adult;
-
-    public Person() {
-
-    }
-
-    public Person(String name, int age) {
-        super();
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public boolean isAdult() {
-        return adult;
-    }
-
-    public void setAdult(boolean adult) {
-        this.adult = adult;
-    }
-
-    @Override
-    public String toString() {
-        return "Person [name=" + name + ", age=" + age + ", adult=" + adult + "]";
-    }
-
-}
