--- process-monitoring-springboot/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-# Packaging
-# quarkus.package.type=fast-jar
-
-infinispan.remote.server-list=localhost:11222
-
-spring.kafka.bootstrap-servers=localhost:9092
-spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
