--- process-monitoring-springboot/src/main/resources/org/kie/kogito/examples/PersonValidationService.drl
@@ -1,28 +1,0 @@
-/**
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.examples
-unit PersonValidationService
-
-import org.kie.kogito.examples.demo.Person;
-
-rule "Is adult"
-when
-    $person: /persons[age > 18]
-then
-    modify($person) { 
-    	setAdult(true) 
-    };
-end
