--- process-monitoring-springboot/src/test/resources/application.properties
@@ -1,4 +1,0 @@
-# Kafka
-spring.kafka.bootstrap-servers=127.0.0.1:9092
-spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
